Move PDataTFStream into separate file
